feat(SPV-1241): change totpClient name to totpAPI; move sign transactions to transactionSigner struct
