+++ internal/api/v1/user/totp/totp.go
@@ -22,18 +22,18 @@ const (
 	DefaultDigits uint = 2
 )
 
-// Client handles TOTP generation and validation.
-type Client struct {
+// API handles TOTP generation and validation.
+type API struct {
 	xPriv *bip32.ExtendedKey
 }
 
 // New creates a new TOTP WalletClient.
-func New(xPriv *bip32.ExtendedKey) *Client {
-	return &Client{xPriv: xPriv}
+func New(xPriv *bip32.ExtendedKey) *API {
+	return &API{xPriv: xPriv}
 }
 
 // GenerateTotpForContact generates a time-based one-time password (TOTP) for a contact.
-func (b *Client) GenerateTotpForContact(contact *models.Contact, period, digits uint) (string, error) {
+func (b *API) GenerateTotpForContact(contact *models.Contact, period, digits uint) (string, error) {
 	sharedSecret, err := b.makeSharedSecret(contact)
 	if err != nil {
 		return "", fmt.Errorf("generateTotpForContact: error when making shared: %w", err)
@@ -48,7 +48,7 @@ func (b *Client) GenerateTotpForContact(contact *models.Contact, period, digits 
 }
 
 // ValidateTotpForContact validates a TOTP for a contact.
-func (b *Client) ValidateTotpForContact(contact *models.Contact, passcode, requesterPaymail string, period, digits uint) error {
+func (b *API) ValidateTotpForContact(contact *models.Contact, passcode, requesterPaymail string, period, digits uint) error {
 	sharedSecret, err := b.makeSharedSecret(contact)
 	if err != nil {
 		return fmt.Errorf("ValidateTotpForContact: error when making shared secret: %w", err)
@@ -65,7 +65,7 @@ func (b *Client) ValidateTotpForContact(contact *models.Contact, passcode, reque
 	return nil
 }
 
-func (b *Client) makeSharedSecret(contact *models.Contact) ([]byte, error) {
+func (b *API) makeSharedSecret(contact *models.Contact) ([]byte, error) {
 	privKey, pubKey, err := b.getSharedSecretFactors(contact)
 	if err != nil {
 		return nil, fmt.Errorf("makeSharedSecret: error when getting shared secret factors: %w", err)
@@ -75,7 +75,7 @@ func (b *Client) makeSharedSecret(contact *models.Contact) ([]byte, error) {
 	return x.Bytes(), nil
 }
 
-func (b *Client) getSharedSecretFactors(contact *models.Contact) (*ec.PrivateKey, *ec.PublicKey, error) {
+func (b *API) getSharedSecretFactors(contact *models.Contact) (*ec.PrivateKey, *ec.PublicKey, error) {
 	if b.xPriv == nil {
 		return nil, nil, errors.ErrMissingXpriv
 	}
